requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
PyYAML
joblib
pathlib